feat: update to react 19 (#361)
[Storybook
Link](https://wandelbotsgmbh.github.io/wandelbots-js-react-components/overview.html)

This updates the dependencies to be compatible with react 19. 
Some MUI components had to be switched to the new Grid component
Storybook is also updated to version 9 which had some breaking changes

I removed spring and replaced it with a custom interpolation lib as the
latest update was to buggy.
https://github.com/pmndrs/react-spring/issues/2377